Handle screenshot in aux windows, move screenshot code into browserHostService
This encapsulates the function that we don't want electron consumers to use,
instead of throwing which we did before.
