+++ src/vs/platform/native/electron-main/nativeHostMainService.ts
@@ -690,8 +690,8 @@ export class NativeHostMainService extends Disposable implements INativeHostMain
 
 	//#region Screenshots
 
-	async getScreenshot(windowId: number | undefined): Promise<ArrayBufferLike | undefined> {
-		const window = this.windowById(undefined, windowId);
+	async getScreenshot(windowId: number | undefined, options?: INativeHostOptions): Promise<ArrayBufferLike | undefined> {
+		const window = this.windowById(options?.targetWindowId, windowId);
 		const captured = await window?.win?.webContents.capturePage();
 		return captured?.toJPEG(95);
 	}
