+++ src/vs/workbench/services/host/browser/browserHostService.ts
@@ -16,7 +16,7 @@ import { IWorkspace, IWorkspaceProvider } from '../../../browser/web.api.js';
 import { IFileService } from '../../../../platform/files/common/files.js';
 import { ILabelService, Verbosity } from '../../../../platform/label/common/label.js';
 import { EventType, ModifierKeyEmitter, addDisposableListener, addDisposableThrottledListener, detectFullscreen, disposableWindowInterval, getActiveDocument, getWindowId, onDidRegisterWindow, trackFocus } from '../../../../base/browser/dom.js';
-import { Disposable } from '../../../../base/common/lifecycle.js';
+import { Disposable, DisposableStore, toDisposable } from '../../../../base/common/lifecycle.js';
 import { IBrowserWorkbenchEnvironmentService } from '../../environment/browser/environmentService.js';
 import { memoize } from '../../../../base/common/decorators.js';
 import { parseLineAndColumnAware } from '../../../../base/common/extpath.js';
@@ -40,7 +40,6 @@ import { coalesce } from '../../../../base/common/arrays.js';
 import { mainWindow, isAuxiliaryWindow } from '../../../../base/browser/window.js';
 import { isIOS, isMacintosh } from '../../../../base/common/platform.js';
 import { IUserDataProfilesService } from '../../../../platform/userDataProfile/common/userDataProfile.js';
-import { getScreenshotViaDisplayMedia } from '../../../../base/browser/screenshot.js';
 
 enum HostShutdownReason {
 
@@ -577,8 +576,66 @@ export class BrowserHostService extends Disposable implements IHostService {
 
 	//#region Screenshots
 
-	getScreenshot(): Promise<ArrayBufferLike | undefined> {
-		return getScreenshotViaDisplayMedia();
+	async getScreenshot(): Promise<ArrayBufferLike | undefined> {
+		// Gets a screenshot from the browser. This gets the screenshot via the browser's display
+		// media API which will typically offer a picker of all available screens and windows for
+		// the user to select. Using the video stream provided by the display media API, this will
+		// capture a single frame of the video and convert it to a JPEG image.
+		const store = new DisposableStore();
+
+		// Create a video element to play the captured screen source
+		const video = document.createElement('video');
+		store.add(toDisposable(() => video.remove()));
+		let stream: MediaStream | undefined;
+		try {
+			// Create a stream from the screen source (capture screen without audio)
+			stream = await navigator.mediaDevices.getDisplayMedia({
+				audio: false,
+				video: true
+			});
+
+			// Set the stream as the source of the video element
+			video.srcObject = stream;
+			video.play();
+
+			// Wait for the video to load properly before capturing the screenshot
+			await Promise.all([
+				new Promise<void>(r => store.add(addDisposableListener(video, 'loadedmetadata', () => r()))),
+				new Promise<void>(r => store.add(addDisposableListener(video, 'canplaythrough', () => r())))
+			]);
+
+			const canvas = document.createElement('canvas');
+			canvas.width = video.videoWidth;
+			canvas.height = video.videoHeight;
+
+			const ctx = canvas.getContext('2d');
+			if (!ctx) {
+				return undefined;
+			}
+
+			// Draw the portion of the video (x, y) with the specified width and height
+			ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
+
+			// Convert the canvas to a Blob (JPEG format), use .95 for quality
+			const blob: Blob | null = await new Promise((resolve) => canvas.toBlob((blob) => resolve(blob), 'image/jpeg', 0.95));
+			if (!blob) {
+				throw new Error('Failed to create blob from canvas');
+			}
+
+			// Convert the Blob to an ArrayBuffer
+			return blob.arrayBuffer();
+
+		} catch (error) {
+			console.error('Error taking screenshot:', error);
+			return undefined;
+		} finally {
+			store.dispose();
+			if (stream) {
+				for (const track of stream.getTracks()) {
+					track.stop();
+				}
+			}
+		}
 	}
 
 	//#endregion
